Minor renaming in common dto classes
