+++ common/src/commonMain/kotlin/com.example.auth.common/decode/DecodedValueDto.kt
@@ -1,9 +1,9 @@
-package com.example.auth.common.encode
+package com.example.auth.common.decode
 
 import com.example.auth.common.byte.RawBytesDto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserDecodeRequestDto(
-    val encodedData: RawBytesDto,
+data class DecodedValueDto(
+    val value: RawBytesDto,
 )
