+++ common/src/commonMain/kotlin/com.example.auth.common/decode/UserDecodeRequestDto.kt
@@ -1,0 +1,9 @@
+package com.example.auth.common.decode
+
+import com.example.auth.common.byte.RawBytesDto
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserDecodeRequestDto(
+    val encodedData: RawBytesDto,
+)
